backend/ppttopdf/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-pptx>=0.6.23
gunicorn==21.2.0
werkzeug==3.0.1
